services/pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "personalassistant-services"
version = "0.1.0"
description = "AI powered personal assistant project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "caldav>=2.1.2",
    "duckduckgo-search>=6.4.2",
    "langchain>=1.0.5",
    "langchain-classic>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.3",
    "llama-index>=0.14.8",
    "llama-index-embeddings-ollama>=0.8.4",
    "llama-index-llms-ollama>=0.9.0",
    "llama-index-llms-openai>=0.6.8",
    "llama-index-tools-duckduckgo>=0.4.1",
    "llama-index-vector-stores-qdrant>=0.3.0",
    "psycopg2-binary==2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "qdrant-client>=1.15.1",
]

[tool.hatch.build.targets.wheel]
packages = ["ai_gateway"]

[tool.hatch.build]
include = [
    "ai_gateway/config/agents.yaml",
]

[dependency-groups]
dev = [
    "ruff==0.14.5",
    "pyright==1.1.407",
    "pre-commit>=4.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["ai_gateway"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []
